Break up `merge-coverage-reports.ts`
